requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.2.2+cpu
flask
gunicorn
requests
beautifulsoup4
python-dotenv
langchain
langchain-chroma
langchain-huggingface
langchain-groq
pypdf
sentence-transformers
flask-cors
edge-tts
